@@ -15,8 +15,6 @@ import org.apache.commons.io.FileUtils;
 
 public class ScreenShotMaker {
 
-    private static String FOLDER = "screenshot";
-
     public static void takeErrorScreenShot(WebDriver webDriver) {
         try {
             shot(webDriver);
@@ -27,10 +25,9 @@ public class ScreenShotMaker {
 
     private static void shot(WebDriver webDriver) throws IOException {
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH.mm.ss");
-        long time = new Date().getTime();
-        Path screenShotPath = Paths.get(FOLDER, "screenshot_" + dateFormat.format(time));
-        TakesScreenshot scrShot = ((TakesScreenshot) webDriver);
-        File screenShotFile = scrShot.getScreenshotAs(OutputType.FILE);
+        String FOLDER = "screenshot";
+        Path screenShotPath = Paths.get(FOLDER, "screenshot_" + dateFormat.format(new Date().getTime()));
+        File screenShotFile = ((TakesScreenshot) webDriver).getScreenshotAs(OutputType.FILE);
         FileUtils.copyFile(screenShotFile, new File(screenShotPath + ".png"));
     }
 }
