@@ -3,18 +3,18 @@ package page.pages.addtobasket;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
+import page.common.BasePage;
 
-class AddToBasketElementProvider {
+class AddToBasketElementProvider extends BasePage {
 
-    AddToBasketElementProvider(WebDriver webDriver) {
-        PageFactory.initElements(webDriver, this);
+    public AddToBasketElementProvider(WebDriver webDriver) {
+        super(webDriver);
     }
 
     @FindBy(id = "addToBasket_vjavse_w")
     private WebElement addToBasketButton;
 
-    WebElement getAddToBasketButton() {
+    public WebElement getAddToBasketButton() {
         return addToBasketButton;
     }
 }
